% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_functions.R
\docType{package}
\name{admin_googleAuthR}
\alias{admin_googleAuthR}
\alias{admin_googleAuthR-package}
\title{Admin Data Transfer API
Transfers user data from one user to another.}
\description{
Auto-generated code by googleAuthR::gar_create_api_skeleton
 at 2016-09-03 22:32:10
filename: /Users/mark/dev/R/autoGoogleAPI/googleadmindatatransferv1.auto/R/admin_functions.R
api_json: api_json
}
\details{
Authentication scopes used are:
\itemize{
  \item https://www.googleapis.com/auth/admin.datatransfer
\item https://www.googleapis.com/auth/admin.datatransfer.readonly
}
}

