% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_objects.R
\name{Application}
\alias{Application}
\title{Admin Data Transfer API Objects 
Transfers user data from one user to another.}
\usage{
Application(etag = NULL, id = NULL, name = NULL, transferParams = NULL)
}
\arguments{
\item{etag}{Etag of the resource}

\item{id}{The application's ID}

\item{name}{The application's name}

\item{transferParams}{The list of all possible transfer parameters for this application}
}
\value{
Application object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2016-09-03 22:32:11
filename: /Users/mark/dev/R/autoGoogleAPI/googleadmindatatransferv1.auto/R/admin_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
Application Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The JSON template for an Application resource.
}

