% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_objects.R
\name{ApplicationDataTransfer}
\alias{ApplicationDataTransfer}
\title{ApplicationDataTransfer Object}
\usage{
ApplicationDataTransfer(applicationId = NULL,
  applicationTransferParams = NULL, applicationTransferStatus = NULL)
}
\arguments{
\item{applicationId}{The application's ID}

\item{applicationTransferParams}{The transfer parameters for the application}

\item{applicationTransferStatus}{Current status of transfer for this application}
}
\value{
ApplicationDataTransfer object
}
\description{
ApplicationDataTransfer Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Template to map fields of ApplicationDataTransfer resource.
}
