% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_objects.R
\name{ApplicationsListResponse}
\alias{ApplicationsListResponse}
\title{ApplicationsListResponse Object}
\usage{
ApplicationsListResponse(applications = NULL, etag = NULL,
  nextPageToken = NULL)
}
\arguments{
\item{applications}{List of applications that support data transfer and are also installed for the customer}

\item{etag}{ETag of the resource}

\item{nextPageToken}{Continuation token which will be used to specify next page in list API}
}
\value{
ApplicationsListResponse object
}
\description{
ApplicationsListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Template for a collection of Applications.
}
