% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_objects.R
\name{ApplicationTransferParam}
\alias{ApplicationTransferParam}
\title{ApplicationTransferParam Object}
\usage{
ApplicationTransferParam(key = NULL, value = NULL)
}
\arguments{
\item{key}{The type of the transfer parameter}

\item{value}{The value of the coressponding transfer parameter}
}
\value{
ApplicationTransferParam object
}
\description{
ApplicationTransferParam Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Template for application transfer parameters.
}

