% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_functions.R
\name{datatransfer.applications.list}
\alias{datatransfer.applications.list}
\title{Lists the applications available for data transfer for a customer.}
\usage{
datatransfer.applications.list(customerId = NULL, maxResults = NULL,
  pageToken = NULL)
}
\arguments{
\item{customerId}{Immutable ID of the Google Apps account}

\item{maxResults}{Maximum number of results to return}

\item{pageToken}{Token to specify next page in the list}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/admin.datatransfer
\item https://www.googleapis.com/auth/admin.datatransfer.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/admin.datatransfer, https://www.googleapis.com/auth/admin.datatransfer.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/admin-sdk/data-transfer/}{Google Documentation}
}

