% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_objects.R
\name{DataTransfer}
\alias{DataTransfer}
\title{DataTransfer Object}
\usage{
DataTransfer(applicationDataTransfers = NULL, etag = NULL, id = NULL,
  newOwnerUserId = NULL, oldOwnerUserId = NULL,
  overallTransferStatusCode = NULL, requestTime = NULL)
}
\arguments{
\item{applicationDataTransfers}{List of per application data transfer resources}

\item{etag}{ETag of the resource}

\item{id}{The transfer's ID (Read-only)}

\item{newOwnerUserId}{ID of the user to whom the data is being transfered}

\item{oldOwnerUserId}{ID of the user whose data is being transfered}

\item{overallTransferStatusCode}{Overall transfer status (Read-only)}

\item{requestTime}{The time at which the data transfer was requested (Read-only)}
}
\value{
DataTransfer object
}
\description{
DataTransfer Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The JSON template for a DataTransfer resource.
}
\seealso{
Other DataTransfer functions: \code{\link{datatransfer.transfers.insert}}
}

