% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_functions.R
\name{datatransfer.transfers.get}
\alias{datatransfer.transfers.get}
\title{Retrieves a data transfer request by its resource ID.}
\usage{
datatransfer.transfers.get(dataTransferId)
}
\arguments{
\item{dataTransferId}{ID of the resource to be retrieved}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/admin.datatransfer
\item https://www.googleapis.com/auth/admin.datatransfer.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/admin.datatransfer, https://www.googleapis.com/auth/admin.datatransfer.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/admin-sdk/data-transfer/}{Google Documentation}
}

