% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_functions.R
\name{datatransfer.transfers.insert}
\alias{datatransfer.transfers.insert}
\title{Inserts a data transfer request.}
\usage{
datatransfer.transfers.insert(DataTransfer)
}
\arguments{
\item{DataTransfer}{The \link{DataTransfer} object to pass to this method
#' @importFrom googleAuthR gar_api_generator}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/admin.datatransfer
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/admin.datatransfer)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/admin-sdk/data-transfer/}{Google Documentation}

Other DataTransfer functions: \code{\link{DataTransfer}}
}

