% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_functions.R
\name{datatransfer.transfers.list}
\alias{datatransfer.transfers.list}
\title{Lists the transfers for a customer by source user, destination user, or status.}
\usage{
datatransfer.transfers.list(customerId = NULL, maxResults = NULL,
  newOwnerUserId = NULL, oldOwnerUserId = NULL, pageToken = NULL,
  status = NULL)
}
\arguments{
\item{customerId}{Immutable ID of the Google Apps account}

\item{maxResults}{Maximum number of results to return}

\item{newOwnerUserId}{Destination user's profile ID}

\item{oldOwnerUserId}{Source user's profile ID}

\item{pageToken}{Token to specify the next page in the list}

\item{status}{Status of the transfer}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/admin.datatransfer
\item https://www.googleapis.com/auth/admin.datatransfer.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/admin.datatransfer, https://www.googleapis.com/auth/admin.datatransfer.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/admin-sdk/data-transfer/}{Google Documentation}
}

