% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_objects.R
\name{DataTransfersListResponse}
\alias{DataTransfersListResponse}
\title{DataTransfersListResponse Object}
\usage{
DataTransfersListResponse(dataTransfers = NULL, etag = NULL,
  nextPageToken = NULL)
}
\arguments{
\item{dataTransfers}{List of data transfer requests}

\item{etag}{ETag of the resource}

\item{nextPageToken}{Continuation token which will be used to specify next page in list API}
}
\value{
DataTransfersListResponse object
}
\description{
DataTransfersListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Template for a collection of DataTransfer resources.
}

